Test the export logging functionality and move the test backend
Because:
- We want to ensure the export logging get's hit in the code, so that it
  can be usedin our applications and we know what comes in that
  function.
- Moved the test backend so that it can be used over multiple test files
  rather than just belonging to http_ex_test. Support seems like a
  logical place for it.
