+++ mix.exs
@@ -9,6 +9,7 @@ defmodule HttpEx.MixProject do
       dialyzer: dialyzer_config(),
       docs: docs(),
       elixir: "~> 1.18",
+      elixirc_paths: elixirc_paths(Mix.env()),
       name: "HTTPEx",
       organization: "wuunder",
       package: package(),
@@ -18,6 +19,9 @@ defmodule HttpEx.MixProject do
       version: "0.1.0"
     ]
   end
+
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 
   # Run "mix help compile.app" to learn about applications.
   def application do
